Merge 26.05 into 26.04

+++ storage/ndb/rest-server2/server/src/ttl_purge.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2024, 2025 Hopsworks and/or its affiliates
+ * Copyright (C) 2024, 2026 Hopsworks and/or its affiliates
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -201,7 +201,7 @@ class TTLPurger {
   // minus_sec lets callers encode a past instant (now - minus_sec). The TTL
   // purge scan passes ttl_sec + purge_window to obtain the expiry threshold.
   static Int64 GetNow(unsigned char* encoded_now, bool timestamp,
-                      Uint32 minus_sec = 0);
+                      Uint64 minus_sec = 0);
   bool UpdateLease(const unsigned char* encoded_now);
   bool IsNodeAlive(const unsigned char* encoded_last_active);
   Uint32 AdjustBatchSize(Uint32 curr_batch_size,
